requirements.txt
plotly==5.13.0
streamlit==1.20.0
matplotlib==3.7.1
streamlit-keplergl
numerize==0.12
pillow==9.4.0
plotly

